+++ NEWS.md
@@ -4,6 +4,8 @@
 - Add support for the zfec forward error correction API now
   available in Botan 3
 - Add more Wycheproof tests
+- Add support for loading DSA public and private keys using
+  ``Pubkey::load_dsa`` and ``Privkey::load_dsa``
 
 ## 0.9.2 2023-02-24